adds evaluation for set symbols and contains
